adds a navigation dashboard imports navigation form in to dashboard removes un-needed mui styling
