+++ src/theme/navigationTheme.js
@@ -1,0 +1,8 @@
+import { makeStyles } from '@material-ui/core'
+
+const navigationDashboard = makeStyles((theme) => ({
+  container: {
+    margin: '5%',
+  },
+}))
+export default navigationDashboard
